przeniesienie z HomeContoller do ReservationController; zaczęcie refaktoru w stronę angielskich nazw
