+++ Hotel.Presentation/Views/Reservation/SprawdzDostepnosc.cshtml
@@ -63,7 +63,6 @@
                         
                         <div class="col-md-4">
                             <div class="card mb-4 shadow-sm">
-                                <!-- <img src="~/Images/@@item.RoomImage" class="card-img-top" height="225"> -->
                                 <div class="card-body">
                                     <h5 class="card-title">@Html.DisplayFor(modelItem => pokoj.PokojTyp.NazwaTypuPokoju) </h5>
                                     <h4 class="card-subtitle"> @Html.DisplayFor(modelItem => pokoj.CenaZaNoc) zł/noc</h4>
@@ -73,7 +72,7 @@
                                     <div class="d-flex justify-content-between align-items-center">
 
                                         <a type="button" class="btn btn-sm btn-outline-secondary" asp-action="SzczegolyPokoju" asp-route-id="@pokoj.Id">Szczegóły</a>
-                                        <input type="submit" value="Zarezerwuj" asp-action="UtworzRezerwacje" asp-controller="Home" class="btn btn-primary" asp-route-id="@pokoj.Id" />
+                                        <input type="submit" value="Zarezerwuj" asp-action="UtworzRezerwacje" asp-controller="Reservation" class="btn btn-primary" asp-route-id="@pokoj.Id" />
 
 
                                     </div>
