--- Hotel.Presentation/Views/Rezerwacja/UtworzRezerwacje1.cshtml
@@ -1,60 +1,0 @@
-﻿
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <title>Formularz z Datami i Liczbą</title>
-
-    <!-- Dodaj linki do arkuszy stylów Bootstrap i Bootstrap-datepicker -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" />
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datepicker/1.9.0/css/bootstrap-datepicker.min.css" />
-</head>
-<body>
-
-    <div class="container mt-5">
-        <h2>Podaj dane rezerwacji</h2>
-
-        <form action="/Rezerwacja/UtworzRezerwacje" method="post">
-
-            <div class="input-group date" data-provide="datepicker">
-                <label for="date1">Data od </label>
-                <input type="text" class="form-control" id="date1" name="dataOd" required />
-            </div>
-
-            <div class="input-group date" data-provide="datepicker">
-                <label for="date2">Data do </label>
-                <input type="text" class="form-control" id="date2" name="dataDo" required />
-            </div>
-
-            <div class="form-group">
-                <label for="number">Liczba osob</label>
-                <input type="number" class="form-control" id="number" name="iloscOsob" required />
-            </div>
-
-            <div class="form-group mb-3">
-                <input type="submit" value="Potwierdz rezerwacje" class="btn btn-primary">
-
-            </div>
-        </form>
-    </div>
-
-    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/@@popperjs/core@2.9.2/dist/umd/popper.min.js"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datepicker/1.9.0/js/bootstrap-datepicker.min.js"></script>
-
-    <script>
-        // Inicjalizuj Datepickery
-        $(document).ready(function () {
-            $('.input-group.date').datepicker({
-                format: 'yyyy-mm-dd',
-                autoclose: true,
-                todayHighlight: true
-            });
-        });
-    </script>
-
-</body>
-</html>
