+++ Hotel.Presentation/Views/Shared/_Layout.cshtml
@@ -12,25 +12,18 @@
     <header>
         <nav class="navbar navbar-expand-sm navbar-toggleable-sm navbar-light bg-white border-bottom box-shadow mb-3">
             <div class="container-fluid">
-                <a class="navbar-brand" asp-area="" asp-controller="Home" asp-action="PulpitRezerwacji">System zarządzania hotelem</a>
+                <a class="navbar-brand" asp-area="" asp-controller="Reservation" asp-action="PulpitRezerwacji">System zarządzania hotelem</a>
                 <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target=".navbar-collapse" aria-controls="navbarSupportedContent"
                         aria-expanded="false" aria-label="Toggle navigation">
                     <span class="navbar-toggler-icon"></span>
                 </button>
                 <div class="navbar-collapse collapse d-sm-inline-flex justify-content-between">
                     <ul class="navbar-nav flex-grow-1">
-                      <!---  <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Index">Home</a>
+                        <li class="nav-item">
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Reservation" asp-action="SprawdzDostepnosc">Sprawdź dostępność</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
-                        </li>
-                        --->
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="SprawdzDostepnosc">Sprawdź dostępność</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="SzczegolyRezerwacji">Sprawdź szczegóły rezerwacji</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Reservation" asp-action="SzczegolyRezerwacji">Sprawdź szczegóły rezerwacji</a>
                         </li>
                     </ul>
                 </div>
